backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0

SQLAlchemy==2.0.48
PyMySQL==1.1.2

python-dotenv==1.2.2
python-jose==3.5.0

passlib==1.7.4
# passlib 1.7.4 es compatible con bcrypt 3.x (bcrypt 4.x rompe __about__)
bcrypt==3.2.2

# necesario para Form/File (multipart/form-data)
python-multipart==0.0.22